DAOS-9917 Auto configure server side YMLs
Removed server and client config .yaml files from
the images directory.  Config files are no longer
added in the images and expected to be modified when
the instances start.

Config files are now templates in terraform.  The
templates are rendered and added to metadata values
for each instance.  Then a startup script is used to
get the content of the metadata values and save it to
the config doas_*.yml files.

This allows settings such as nr_hugepages, targets, scm_size
in the config files to be set properly based on the
terraform variables.

The daos_server terraform module now has an output value
named access_points that contains a list of the access points
needed for the daos_agent.yml and daos_control.yml on the
clients.

Signed-off-by: Mark A. Olson <mark.a.olson@intel.com>