DAOSGCP-139 Misc fixes after further testing
Added set -o pipefail to bash scripts

Increased timeout for certificate deployment because with 32
clients and 10 servers it was timing out. The timeout was doubled
from 5 mins to 10 mins. That's way too long. Need to improve
cert deployment solution in a future ticket.

Updated cloudshell_urls.sh so that it updates "git clone"
commands with URLs to match the repo used for the cloudshell
tutorial.

Signed-off-by: Mark A. Olson <mark.a.olson@intel.com>