DAOSGCP-139 Update for DAOS v2.2.0 and IO500-SC22 (#67)
Updates to support DAOS v2.2.0 and IO500-SC22

Updated the development.md document to include workflow steps
for development and releases.

Use consistent term "Destroy" for destroying all resources
with the `terraform destroy` command.

Added set -o pipefail to bash scripts

Increased timeout for certificate deployment because with 32
clients and 10 servers it was timing out. The timeout was doubled
from 5 mins to 10 mins. That's way too long. Need to improve
cert deployment solution in a future ticket.

Updated cloudshell_urls.sh so that it updates "git clone"
commands with URLs to match the repo used for the cloudshell
tutorial.

Signed-off-by: Mark Olson <115657904+mark-olson@users.noreply.github.com>